(feat!2) solution 61, still strange
realised that the final condition was overconstraining.
there need not be just one number of each type in the sequence, a number
could be classifed as two polygonal types. surprisingly relaxing the condition
yields many such sequences while the problem statement mentions the presence of
a unique sequence of length 6.